all: move main transaction pool into a subpool
